add assignment scoring and add change pass featcher

+++ OnlineEducationPlatform.Web/Models/AssignmentSubmission.cs
@@ -6,7 +6,7 @@
         public string StudentId { get; set; }
         public string FilePath { get; set; }
         public DateTime SubmittedAt { get; set; }
-        public decimal? Score { get; set; }
+        public int? Score { get; set; }
         public Assignment Assignment { get; set; }
         public Student Student { get; set; }
     }
